requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Deprecated==1.2.13
Django==4.1.7
django-oauth-toolkit==2.2.0
django-phonenumber-field==7.0.2
django-redis==5.2.0
django-rest-knox==4.2.0
django-restframework==0.0.1
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
hashring==1.5.1
idna==3.4
jwcrypto==1.4.2
memcache==0.12.0
oauthlib==3.2.2
passlib==1.7.4
phonenumbers==8.13.7
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
pymemcache==4.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.4
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
ua-parser==0.16.1
urllib3==1.26.15
user-agents==2.2.0
wrapt==1.15.0
